+++ src/core.rs
@@ -1,3 +1,2 @@
-pub mod loads;
 pub mod macros;
 pub mod metadata;
