+++ src/core/macros.rs
@@ -1,5 +1,2 @@
 pub mod dumps;
-
 pub mod loads;
-
-pub mod pretty;
