+++ src/lib.rs
@@ -19,20 +19,7 @@ mod toml_rs {
 
     #[pymodule_export]
     use crate::document::TOMLDocument;
-    use crate::{
-        v1::{
-            dumps::{python_to_toml_v1, validate_inline_paths_v1},
-            loads::toml_to_python_v1,
-            metadata::{extract_metadata_v1, to_python_v1},
-            pretty::PrettyV1,
-        },
-        v1_1::{
-            dumps::{python_to_toml, validate_inline_paths},
-            loads::toml_to_python,
-            metadata::{extract_metadata, to_python},
-            pretty::Pretty,
-        },
-    };
+
     #[pymodule_export]
     const _VERSION: &str = env!("CARGO_PKG_VERSION");
 
@@ -61,7 +48,7 @@ mod toml_rs {
                     ))
                 })?;
 
-                let toml = toml_to_python_v1(
+                let toml = crate::v1::loads::toml_to_python(
                     py,
                     &Spanned::new(parsed.span(), DeValue::Table(parsed.into_inner())),
                     parse_float,
@@ -84,7 +71,7 @@ mod toml_rs {
                     ))
                 })?;
 
-                let toml = toml_to_python(
+                let toml = crate::v1_1::loads::toml_to_python(
                     py,
                     &Spanned::new(parsed.span(), DeValue::Table(parsed.into_inner())),
                     parse_float,
@@ -112,24 +99,34 @@ mod toml_rs {
             "1.0.0" => {
                 use toml_edit_v1::{DocumentMut, Item::Table, visit_mut::VisitMut};
 
+                use crate::v1::{
+                    dumps::{python_to_toml, validate_inline_paths},
+                    pretty::Pretty,
+                };
+
                 let mut doc = DocumentMut::new();
 
-                if let Table(table) = python_to_toml_v1(py, obj, inline_tables.as_ref())? {
+                if let Table(table) = python_to_toml(py, obj, inline_tables.as_ref())? {
                     *doc.as_table_mut() = table;
                 }
 
                 if let Some(ref paths) = inline_tables {
-                    validate_inline_paths_v1(doc.as_item(), paths)?;
+                    validate_inline_paths(doc.as_item(), paths)?;
                 }
 
                 if pretty {
-                    PrettyV1::new(inline_tables.is_none()).visit_document_mut(&mut doc);
+                    Pretty::new(inline_tables.is_none()).visit_document_mut(&mut doc);
                 }
 
                 Ok(doc.to_string())
             }
             "1.1.0" => {
                 use toml_edit::{DocumentMut, Item::Table, visit_mut::VisitMut};
+
+                use crate::v1_1::{
+                    dumps::{python_to_toml, validate_inline_paths},
+                    pretty::Pretty,
+                };
 
                 let mut doc = DocumentMut::new();
 
@@ -163,6 +160,8 @@ mod toml_rs {
             "1.0.0" => {
                 use toml_v1::de::{DeTable, DeValue};
 
+                use crate::v1::metadata::{extract_metadata, to_python};
+
                 let parsed = DeTable::parse(toml_string).map_err(|err| {
                     TOMLDecodeError::new_err((
                         err.to_string(),
@@ -171,11 +170,11 @@ mod toml_rs {
                     ))
                 })?;
 
-                let meta = extract_metadata_v1(py, &parsed, toml_string)?;
+                let meta = extract_metadata(py, &parsed, toml_string)?;
 
                 let span = parsed.span();
                 let inner = parsed.into_inner();
-                let value = to_python_v1(py, &DeValue::Table(inner), span, toml_string)?;
+                let value = to_python(py, &DeValue::Table(inner), span, toml_string)?;
 
                 let doc = Py::new(
                     py,
@@ -189,6 +188,8 @@ mod toml_rs {
             }
             "1.1.0" => {
                 use toml::de::{DeTable, DeValue};
+
+                use crate::v1_1::metadata::{extract_metadata, to_python};
 
                 let parsed = DeTable::parse(toml_string).map_err(|err| {
                     TOMLDecodeError::new_err((
