+++ src/v1/dumps.rs
@@ -7,8 +7,8 @@ use toml_edit_v1::{Array, InlineTable, Item, Offset, Table, Value};
 use crate::{impl_dumps, to_toml_v1, toml_dt_v1};
 
 impl_dumps!(
-    validate_inline_paths_v1,
-    python_to_toml_v1,
+    validate_inline_paths,
+    python_to_toml,
     to_toml_v1,
     toml_dt_v1
 );
