+++ src/v1_1/loads.rs
@@ -1,19 +1,147 @@
-use pyo3::prelude::*;
-use toml::{Spanned, de::DeValue};
+use num_bigint::BigInt;
+use pyo3::{
+    IntoPyObjectExt,
+    exceptions::PyValueError,
+    prelude::*,
+    types::{PyDate, PyDelta, PyDict, PyList, PyTime, PyTzInfo},
+};
+use toml::{Spanned, de::DeValue, value::Offset};
 
-use crate::{create_py_datetime, impl_loads};
+use crate::{create_py_datetime, error::TomlError, parse_int, toml_rs::TOMLDecodeError};
 
 pub fn toml_to_python<'py>(
     py: Python<'py>,
-    value: &Spanned<DeValue<'_>>,
+    de_value: &Spanned<DeValue<'_>>,
     parse_float: &Bound<'py, PyAny>,
     doc: &str,
 ) -> PyResult<Bound<'py, PyAny>> {
-    to_python(py, value, parse_float, doc)
+    to_python(py, de_value, parse_float, doc)
 }
 
-impl_loads!(
-    create_py_datetime,
-    |time: &toml::value::Time| time.second.unwrap_or(0),
-    |time: &toml::value::Time| time.nanosecond.unwrap_or(0) / 1000
-);
+fn to_python<'py>(
+    py: Python<'py>,
+    de_value: &Spanned<DeValue<'_>>,
+    parse_float: &Bound<'py, PyAny>,
+    doc: &str,
+) -> PyResult<Bound<'py, PyAny>> {
+    let value = de_value.as_ref();
+    let span = de_value.span();
+
+    match value {
+        DeValue::String(str) => str.into_bound_py_any(py),
+        DeValue::Integer(int) => {
+            let bytes = int.as_str().as_bytes();
+            let radix = int.radix();
+
+            let parse_options = lexical_core::ParseIntegerOptions::new();
+
+            if let Ok(int_64) = parse_int!(i64, bytes, &parse_options, radix) {
+                return int_64.into_bound_py_any(py);
+            }
+
+            if let Some(big_int) = BigInt::parse_bytes(bytes, radix) {
+                return big_int.into_bound_py_any(py);
+            }
+
+            let error_start = span.start;
+            let mut err = TomlError::custom(
+                format!(
+                    "invalid integer '{}'",
+                    &doc[span.start..span.end.min(doc.len())]
+                ),
+                Some(span),
+            );
+            err.set_input(Some(doc));
+
+            Err(TOMLDecodeError::new_err((
+                err.to_string(),
+                doc.to_string(),
+                error_start,
+            )))
+        }
+        DeValue::Float(float) => {
+            let float_str = float.as_str();
+
+            let py_call = parse_float.call1((float_str,))?;
+
+            // https://github.com/hukkin/tomli/blob/2.4.1/src/tomli/_parser.py#L789
+            if py_call.is_instance_of::<PyDict>() || py_call.is_instance_of::<PyList>() {
+                // https://github.com/hukkin/tomli/blob/2.4.1/src/tomli/_parser.py#L790
+                return Err(PyValueError::new_err(
+                    "parse_float must not return dicts or lists",
+                ));
+            }
+
+            Ok(py_call)
+        }
+        DeValue::Boolean(bool) => bool.into_bound_py_any(py),
+        DeValue::Datetime(datetime) => match (datetime.date, datetime.time, datetime.offset) {
+            (Some(date), Some(time), Some(offset)) => {
+                let py_tzinfo = create_timezone_from_offset(py, offset)?;
+                let tzinfo = Some(&py_tzinfo);
+                Ok(create_py_datetime!(py, date, time, tzinfo)?.into_any())
+            }
+            (Some(date), Some(time), None) => {
+                Ok(create_py_datetime!(py, date, time, None)?.into_any())
+            }
+            (Some(date), None, None) => {
+                let py_date = PyDate::new(py, i32::from(date.year), date.month, date.day)?;
+                Ok(py_date.into_any())
+            }
+            (None, Some(time), None) => {
+                let py_time = PyTime::new(
+                    py,
+                    time.hour,
+                    time.minute,
+                    time.second.unwrap_or(0),
+                    time.nanosecond.unwrap_or(0) / 1000,
+                    None,
+                )?;
+
+                Ok(py_time.into_any())
+            }
+
+            _ => unreachable!(),
+        },
+
+        DeValue::Array(array) => {
+            if array.is_empty() {
+                return Ok(PyList::empty(py).into_any());
+            }
+
+            let py_list = PyList::empty(py);
+            for item in array {
+                py_list.append(to_python(py, item, parse_float, doc)?)?;
+            }
+            Ok(py_list.into_any())
+        }
+
+        DeValue::Table(table) => {
+            if table.is_empty() {
+                return Ok(PyDict::new(py).into_any());
+            }
+
+            let py_dict = PyDict::new(py);
+            for (key, value) in table {
+                py_dict.set_item(key.as_ref(), to_python(py, value, parse_float, doc)?)?;
+            }
+            Ok(py_dict.into_any())
+        }
+    }
+}
+
+#[inline]
+pub fn create_timezone_from_offset(py: Python, offset: Offset) -> PyResult<Bound<PyTzInfo>> {
+    const SECS_IN_DAY: i32 = 86_400;
+
+    match offset {
+        Offset::Z => PyTzInfo::utc(py).map(Borrowed::to_owned),
+        Offset::Custom { minutes } => {
+            let seconds = i32::from(minutes) * 60;
+            let days = seconds.div_euclid(SECS_IN_DAY);
+            let seconds = seconds.rem_euclid(SECS_IN_DAY);
+            let py_delta = PyDelta::new(py, days, seconds, 0, false)?;
+            PyTzInfo::fixed_offset(py, py_delta)
+        }
+    }
+}
